+++ windowfunctions.h
@@ -28,9 +28,9 @@ namespace KeyFinder {
 
   class WindowFunction {
   public:
-    float window(temporal_window_t windowType, int sample, int width) const;
-    float gaussianWindow(int sample, int width, float sigma) const;
-    std::vector<float> convolve(const std::vector<float>& input, const std::vector<float>& window) const;
+    double window(temporal_window_t windowType, int sample, int width) const;
+    double gaussianWindow(int sample, int width, double sigma) const;
+    std::vector<double> convolve(const std::vector<double>& input, const std::vector<double>& window) const;
   };
 
 }
